requirements-dev.txt
black
codespell
flake8
mypy
pytest
pytest-cov
pytest-mock
pytest-django
pytest-playwright
freezegun
twine
wheel
django-extensions
-e .
